committing after long time, don't know what I have done. Project is working.

+++ CimpleChat.Test/GetNextId.cs
@@ -10,41 +10,41 @@ namespace CimpleChat.Test
         public void TestSingleUserId()
         {
             IGetNextId nextId = Helper.GetRequiredService<IGetNextId>();
-            int id = nextId.GetUserId();
+            long id = nextId.GetUserId();
             Assert.True(id > 0);
-            Assert.IsType<int>(id);
+            Assert.IsType<long>(id);
         }
 
         [Fact]
         public void TestSingleChannelId()
         {
             IGetNextId nextId = Helper.GetRequiredService<IGetNextId>();
-            int id = nextId.GetChannelId();
+            long id = nextId.GetChannelId();
             Assert.True(id > 0);
-            Assert.IsType<int>(id);
+            Assert.IsType<long>(id);
         }
 
         [Fact]
         public void TestSingleMessageId()
         {
             IGetNextId nextId = Helper.GetRequiredService<IGetNextId>();
-            int id = nextId.GetMessageId();
+            long id = nextId.GetMessageId();
             Assert.False(id <= 0);
-            Assert.IsType<int>(id);
+            Assert.IsType<long>(id);
         }
 
         [Fact]
         public void TestMultipleUserId()
         {
             IGetNextId nextId = Helper.GetRequiredService<IGetNextId>();
-            List<int> ids = new List<int>();
+            List<long> ids = new List<long>();
             for(int i = 0; i < 255; i++)
             {
-                int id = nextId.GetUserId();
+                long id = nextId.GetUserId();
                 ids.Add(id);
             }
 
-            List<int> d = ids.Distinct().ToList<int>();
+            List<long> d = ids.Distinct().ToList<long>();
             
             Assert.True(ids.Count == d.Count);
         }
@@ -53,14 +53,14 @@ namespace CimpleChat.Test
         public void TestMultipleChannelId()
         {
             IGetNextId nextId = Helper.GetRequiredService<IGetNextId>();
-            List<int> ids = new List<int>();
+            List<long> ids = new List<long>();
             for (int i = 0; i < 255; i++)
             {
-                int id = nextId.GetChannelId();
+                long id = nextId.GetChannelId();
                 ids.Add(id);
             }
 
-            List<int> d = ids.Distinct().ToList<int>();
+            List<long> d = ids.Distinct().ToList<long>();
 
             Assert.True(ids.Count == d.Count);
         }
@@ -69,14 +69,14 @@ namespace CimpleChat.Test
         public void TestMultipleMessageId()
         {
             IGetNextId nextId = Helper.GetRequiredService<IGetNextId>();
-            List<int> ids = new List<int>();
+            List<long> ids = new List<long>();
             for (int i = 0; i < 255; i++)
             {
-                int id = nextId.GetMessageId();
+                long id = nextId.GetMessageId();
                 ids.Add(id);
             }
 
-            List<int> d = ids.Distinct().ToList<int>();
+            List<long> d = ids.Distinct().ToList<long>();
 
             Assert.True(ids.Count == d.Count);
         }
