+++ CimpleChat/Services/IGetNextId.cs
@@ -2,8 +2,8 @@
 {
     public interface IGetNextId
     {
-        public int GetUserId();
-        public int GetChannelId();
-        public int GetMessageId();
+        public long GetUserId();
+        public long GetChannelId();
+        public long GetMessageId();
     }
 }
